Setup for the first big run
ran it on AWS with first 1000, then 5000 then 2500 executor.queueSize limit. The
higher limits caused input/output errors on /s3mnt, with goofys, or the
container could not be started at all, with a werid error: error while
creating mount source path '/s3mnt': mkdir /s3mnt, file exists.

Possible workarounds/solutions:
- Limit job submission rate via executor.submitRateLimit, eg '100/2min',
100 jobs every 2 minutes.
- Increase temp space on VM, and copy the files for processing there.
- umount/mount bucket, make multiple tests if IO works, and sleep between
attempts.

Also, there were some errors for out of memory during the last step of
compression, thus zstd needs higher memory for the next run.

A snapshot of the status:
[52/c0d68f] process > CloudDiamond:download_sra (ERR3483308)           [100%] 43350 of 43350, cached: 43243, failed: 85 ✔
[52/9403b4] process > CloudDiamond:convert_sra_to_fastq (SRR14139965)  [ 99%] 43219 of 43265, cached: 1886, failed: 21625
[15/544df9] process > CloudDiamond:join_paired_end_reads (ERR9708723)  [ 85%] 18477 of 21594, cached: 245, failed: 10608
executor >  awsbatch (64563)
[52/c0d68f] process > CloudDiamond:download_sra (ERR3483308)           [100%] 43350 of 43350, cached: 43243, failed: 85 ✔
executor >  awsbatch (82701)
[52/c0d68f] process > CloudDiamond:download_sra (ERR3483308)          [100%] 43350 of 43350, cached: 43243, failed: 85 ✔
[87/66a287] process > CloudDiamond:convert_sra_to_fastq (SRR12712406) [100%] 43265 of 43265, cached: 1886, failed: 21628 ✔
[f0/2212f8] process > CloudDiamond:join_paired_end_reads (ERR9661576) [100%] 21637 of 21637, cached: 245, failed: 10752 ✔
[8e/b1c905] process > CloudDiamond:diamond_fastqs (SRR12229366)       [ 95%] 10375 of 10885, cached: 7, failed: 1430
[b3/11a339] process > CloudDiamond:compress_outputs (SRR12229366)     [ 99%] 8937 of 8945, failed: 3072

A problem was also that the download_sra step was the first to be
executed, thus flooding the bucket. Possible workarounds would be to
create multiple queues for each step, and set the later steps to
a higher priority. Alternatively, the workflow of GEMMaker could be
adopted.