Added creation of best hits per read
Added rules and scripts to filter the diamondn files such that only the
best hit - based on bitscore - for each query read is stored.

The same hits histograms are then computed for these filtered files.

This was implemente3d as extra rules so as not to re-run the already
existing analyses, since Snakemake will only run the rules for
non-existing files.